Refactor Swift package to include app extension code in a separate library
