+++ SwiftPackage/Sources/BridgeClientExtension/Data Archiving/ArchiveBuilder.swift
