+++ SwiftPackage/Sources/BridgeClientExtension/Data Archiving/AssessmentResultArchive.swift
