+++ SwiftPackage/Sources/BridgeClientExtension/File Uploading/StudyDataUploadAPI.swift
