+++ SwiftPackage/Sources/BridgeClientExtension/KotlinBase+ClientData.swift
