+++ SwiftPackage/Sources/BridgeClientExtension/Model/PlatformConfigImpl.swift
