+++ SwiftPackage/Sources/BridgeClientExtension/Model/ValidationError.swift
