+++ SwiftPackage/Sources/BridgeClientUI/Assessment Handling/SageResearchTaskDelegate.swift
@@ -32,6 +32,7 @@
 
 import Foundation
 import BridgeClient
+import BridgeClientExtension
 import JsonModel
 import Research
 import ResearchUI
