+++ SwiftPackage/Tests/BridgeClientExtensionTests/Resources/cat.jpg
