+++ SwiftPackage/Tests/BridgeClientExtensionTests/Resources/failed-upload-request-response.json
