+++ SwiftPackage/Tests/BridgeClientExtensionTests/Resources/sd-upload-request-success.json
