+++ SwiftPackage/Tests/BridgeClientExtensionTests/Resources/upload-complete-response.json
